+++ src/Casts/Prices.php
@@ -2,7 +2,7 @@
 
 namespace Bazar\Casts;
 
-use Bazar\Support\Attributes\Prices as PricesBag;
+use Bazar\Support\Bags\Prices as PricesBag;
 use Illuminate\Contracts\Database\Eloquent\CastsAttributes;
 
 class Prices implements CastsAttributes
@@ -14,7 +14,7 @@ class Prices implements CastsAttributes
      * @param  string  $key
      * @param  mixed  $value
      * @param  array  $attributes
-     * @return \Bazar\Support\Attributes\Prices
+     * @return \Bazar\Support\Bags\Prices
      */
     public function get($model, string $key, $value, array $attributes): PricesBag
     {
@@ -28,7 +28,7 @@ class Prices implements CastsAttributes
      *
      * @param  \Illuminate\Database\Eloquent\Model  $model
      * @param  string  $key
-     * @param  \Bazar\Support\Attributes\Prices  $value
+     * @param  \Bazar\Support\Bags\Prices  $value
      * @param  array  $attributes
      * @return string
      */
