+++ src/Support/Bags/Bag.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Bazar\Support\Attributes;
+namespace Bazar\Support\Bags;
 
 use ArrayAccess;
 use ArrayIterator;
