+++ src/Support/Bags/Price.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Bazar\Support\Attributes;
+namespace Bazar\Support\Bags;
 
 use Illuminate\Support\Str;
